+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog for AWS SDK for JavaScript
-<!--LATEST=2.879.0-->
+<!--LATEST=2.880.0-->
 <!--ENTRYINSERT-->
+
+## 2.880.0
+* feature: Appflow: Added destination properties for Zendesk.
+* feature: AuditManager: AWS Audit Manager has updated the GetAssessment API operation to include a new response field called userRole. The userRole field indicates the role information and IAM ARN of the API caller.
+* feature: MediaLive: MediaLive now support HTML5 Motion Graphics overlay
 
 ## 2.879.0
 * bugfix: s3: handle 'continue' event only once